Moved BaseClass dependent unit tests into 3 seperate classes for:
- BaseClass
- SpecializationClass
- Background
