--- RPGManager_development/test/unit_test/character_sheet_unit_tests/BaseClassDependentUnitTests.java
@@ -1,123 +1,0 @@
-package unit_test.character_sheet_unit_tests;
-
-import static org.junit.Assert.assertEquals;
-
-import java.security.InvalidParameterException;
-
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-
-import rpg_database.character_sheet.Background;
-import rpg_database.character_sheet.BaseClasses;
-import rpg_database.character_sheet.CharacterSheet;
-import rpg_database.character_sheet.Fields;
-import rpg_database.character_sheet.SpecializationClasses;
-import rpg_database.character_sheet.exceptions.InvalidCharacterClassException;
-
-import static unit_test.character_sheet_unit_tests.common.CommonMethods.*;
-
-public class BaseClassDependentUnitTests {
-
-	// test methods
-	@Rule
-	public ExpectedException thrown = ExpectedException.none();
-
-	@Test
-	public void testSetCharacterBaseClassMage() {
-		CharacterSheet characterSheet = new CharacterSheet("CharacterSheet");
-		characterSheet.setData(Fields.BASECLASS, BaseClasses.MAGE);
-		BaseClasses actualClass = characterSheet.getData(Fields.BASECLASS);
-		assertEquals(BaseClasses.MAGE, actualClass);
-	}
-
-	@Test
-	public void testSetCharacterBaseClassRogue() {
-		CharacterSheet characterSheet = new CharacterSheet("CharacterSheet");
-		characterSheet.setData(Fields.BASECLASS, BaseClasses.ROGUE);
-		BaseClasses actualClass = characterSheet.getData(Fields.BASECLASS);
-		assertEquals(BaseClasses.ROGUE, actualClass);
-	}
-
-	@Test
-	public void testSetCharacterSpecClassChampion() {
-		CharacterSheet characterSheet = new CharacterSheet("CharacterSheet");
-		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.CHAMPION);
-		SpecializationClasses actualClass = characterSheet.getData(Fields.SPECIALIZATIONCLASS);
-		assertEquals(SpecializationClasses.CHAMPION, actualClass);
-	}
-
-	@Test
-	public void testSetCharacterSpecClassBerserker() {
-		CharacterSheet characterSheet = new CharacterSheet("CharacterSheet");
-		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.BERSERKER);
-		SpecializationClasses actualClass = characterSheet.getData(Fields.SPECIALIZATIONCLASS);
-		assertEquals(SpecializationClasses.BERSERKER, actualClass);
-	}
-
-	@Test
-	public void expectException_SetCharacterBaseClassFromMageToWarrior_ArcaneWarrior() {
-		expectExceptionWithMessage(InvalidCharacterClassException.class, "Warrior is not a base class of Arcane Warrior");
-		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.MAGE, SpecializationClasses.ARCANE_WARRIOR);
-		characterSheet.setData(Fields.BASECLASS, BaseClasses.WARRIOR);
-	}
-
-	@Test
-	public void expectException_SetCharacterBaseClassFromWarriorToMage_Berserker() {
-		expectExceptionWithMessage(InvalidCharacterClassException.class, "Mage is not a base class of Berserker");
-		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.WARRIOR, SpecializationClasses.BERSERKER);
-		characterSheet.setData(Fields.BASECLASS, BaseClasses.MAGE);
-	}
-
-	@Test
-	public void expectException_SetCharacterSpecializationClassFromArcaneWarriorToAssassin_Mage() {
-		expectExceptionWithMessage(InvalidCharacterClassException.class, "Mage is not a base class of Assassin");
-		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.MAGE, SpecializationClasses.ARCANE_WARRIOR);
-		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.ASSASSIN);
-	}
-
-	@Test
-	public void expectException_SetCharacterSpecializationClassFromBerserkerToKeeper_Warrior() {
-		expectExceptionWithMessage(InvalidCharacterClassException.class, "Warrior is not a base class of Keeper");
-		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.WARRIOR, SpecializationClasses.BERSERKER);
-		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.KEEPER);
-	}
-
-	@Test
-	public void expectException_SetCharacterBackgroundMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class,
-				"class java.lang.String value is not an instance of class rpg_database.character_sheet.Background");
-		final String malformedInput = "MALFORMED INPUT";
-		new CharacterSheet("CharacterSheet").setData(Fields.BACKGROUND, malformedInput);
-	}
-
-	@Test
-	public void expectException_SetInvalidBackgroundApostateForClassWarrior() {
-		expectExceptionWithMessage(InvalidCharacterClassException.class, "Apostate is not a Warrior background!");
-		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.WARRIOR, SpecializationClasses.BERSERKER);
-		characterSheet.setData(Fields.BACKGROUND, Background.APOSTATE);
-	}
-
-	@Test
-	public void expectException_SetInvalidBackgroundFereldanFreemanForClassMage() {
-		expectExceptionWithMessage(InvalidCharacterClassException.class, "Fereldan Freeman is not a Mage background!");
-		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.MAGE, SpecializationClasses.KEEPER);
-		characterSheet.setData(Fields.BACKGROUND, Background.FERELDAN_FREEMAN);
-	}
-
-	@Test
-	public void testSetBackgroundChasindWilder() {
-		CharacterSheet characterSheet = new CharacterSheet("characterSheet");
-		characterSheet.setData(Fields.BACKGROUND, Background.CHASIND_WILDER);
-		Background background = characterSheet.getData(Fields.BACKGROUND);
-		assertEquals(Background.CHASIND_WILDER, background);
-	}
-
-	// private methods
-
-	private void expectExceptionWithMessage(Class<? extends Exception> exceptionClass, String message) {
-		thrown.expect(exceptionClass);
-		thrown.expectMessage(message);
-	}
-
-}
