+++ RPGManager_development/test/unit_test/character_sheet_unit_tests/SpecializationClassUnitTests.java
@@ -1,0 +1,60 @@
+package unit_test.character_sheet_unit_tests;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+
+import rpg_database.character_sheet.BaseClasses;
+import rpg_database.character_sheet.CharacterSheet;
+import rpg_database.character_sheet.Fields;
+import rpg_database.character_sheet.SpecializationClasses;
+import rpg_database.character_sheet.exceptions.InvalidCharacterClassException;
+
+import static unit_test.character_sheet_unit_tests.common.CommonMethods.*;
+
+public class SpecializationClassUnitTests {
+
+	// test methods
+	@Rule
+	public ExpectedException thrown = ExpectedException.none();
+
+	@Test
+	public void testSetCharacterSpecClassChampion() {
+		CharacterSheet characterSheet = new CharacterSheet("CharacterSheet");
+		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.CHAMPION);
+		SpecializationClasses actualClass = characterSheet.getData(Fields.SPECIALIZATIONCLASS);
+		assertEquals(SpecializationClasses.CHAMPION, actualClass);
+	}
+
+	@Test
+	public void testSetCharacterSpecClassBerserker() {
+		CharacterSheet characterSheet = new CharacterSheet("CharacterSheet");
+		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.BERSERKER);
+		SpecializationClasses actualClass = characterSheet.getData(Fields.SPECIALIZATIONCLASS);
+		assertEquals(SpecializationClasses.BERSERKER, actualClass);
+	}
+
+	@Test
+	public void expectException_SetCharacterSpecializationClassFromArcaneWarriorToAssassin_Mage() {
+		expectExceptionWithMessage(InvalidCharacterClassException.class, "Mage is not a base class of Assassin");
+		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.MAGE, SpecializationClasses.ARCANE_WARRIOR);
+		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.ASSASSIN);
+	}
+
+	@Test
+	public void expectException_SetCharacterSpecializationClassFromBerserkerToKeeper_Warrior() {
+		expectExceptionWithMessage(InvalidCharacterClassException.class, "Warrior is not a base class of Keeper");
+		CharacterSheet characterSheet = createCharacterSheetWithCustomClasses(BaseClasses.WARRIOR, SpecializationClasses.BERSERKER);
+		characterSheet.setData(Fields.SPECIALIZATIONCLASS, SpecializationClasses.KEEPER);
+	}
+
+	// private methods
+
+	private void expectExceptionWithMessage(Class<? extends Exception> exceptionClass, String message) {
+		thrown.expect(exceptionClass);
+		thrown.expectMessage(message);
+	}
+
+}
